+++ brathika/src/main/java/in/co/nmsworks/training/week2/day5/question2/Phone.java
@@ -9,7 +9,7 @@ public class Phone implements BatteryPowered{
 
     @Override
     public int batteryAfterUse(float hours) {
-        int remainingBattery = 100 - batteryUsagePerHour() * (int)hours;
-        return  remainingBattery;
+        double remainingBattery = 100 - batteryUsagePerHour() * hours;
+        return  (int) remainingBattery;
     }
 }
